feat(useWallet, useStatusMessage):  - reorganizes the useWallet context, provider and hook, in separated files to avoid conflits with React Fast-Refreshing.  - creates the useStatusMessage hook.
